♻️ Extract event.go from step_options.go
